% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{c.GSMetrop}
\alias{CMM}
\alias{c.GSMetrop}
\title{Combine Several Single Metropolis Objects Into One CMM objects}
\usage{
\method{c}{GSMetrop}(...)
}
\arguments{
\item{\dots}{a list of objects of class \code{GSMetrop}}
}
\value{
an object of class "CMM".  If only one object is passed,
the function returns the input object without transformation.
}
\description{
Combine Several Single Metropolis Objects Into One CMM objects
}
\details{
\code{c.GSMetrop} transforms several objects of class
\code{GSMetrop} into an object of class \code{CMM} (combined
multiple Metropolis).
}
\examples{
#############################################
##
## First start with the same examples as in
## the help page of GeneralSingleMetropolis
##
## Generates a dataset:
x <- rnorm(100)
y <- 1+ 2*x + rnorm(100, sd = 0.1)

## Linear regression: estimate a, b, sigma in
## y = a + b*x + epsilon
## epsilon ~ dnorm(0, sigma)

## Log-posterior
logp <- function(par, lidat, ctrl)
{
   ## Current values of the parameters
   a <- par$a
   b <- par$b
   sigma <- exp(par$lsigma)

   ## log-Prior:
   lprior <- dnorm(a, mean=0, sd=100, log=TRUE)
   lprior <- lprior + dnorm(b, mean=0, sd=100, log=TRUE)
   ## uniform prior on sigma

   ## log-likelihood
   llike <- sum(dnorm(lidat$y, mean=lidat$x*b+a, sd=sigma, log=TRUE))

   ## log-posterior
   return(llike+lprior)
}

## Elements required for the fit
parinit <- list(a=0, b=1, lsigma=1) ## initial values
listUpdating <- list(a="sis", b="sis", lsigma="sis") ## all elements are updated with sis
lidat <- list(x=x, y=y)
lipum <- list(a=0.1, b=0.1, lsigma=1) ## test standard deviation of the proposals

\dontrun{
#############################################
##
## Then demonstrate the class CMM
##
## Three runs
gsm1 <- GeneralSingleMetropolis(parinit, lidat, logp, lipum,
                                nrepet=5000, saveResults=FALSE, nburnin=500)
gsm2 <- GeneralSingleMetropolis(parinit, lidat, logp, lipum,
                                nrepet=5000, saveResults=FALSE, nburnin=500)
gsm3 <- GeneralSingleMetropolis(parinit, lidat, logp, lipum,
                                nrepet=5000, saveResults=FALSE, nburnin=500)

## Combine these three runs
ae <- c(gsm1, gsm2, gsm3)

## Show various information about these models
ae
summary(ae)
plot(ae)

## Extraction features:
ae[1:2]
ae[2]

## work with the package coda:
library(coda)
raftery.diag(tocoda(gsm1))

gelman.diag(tocoda(ae))
}
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}}
}

