% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(x, loglikelihood)
}
\arguments{
\item{x}{an object of class \code{CMM} or \code{GSMetrop}}

\item{loglikelihood}{a function for the calculation of the log
likelihood for the current model. This function must rely only on
the data provided in \code{lidat}. This function must have three
parameters: (i) \code{par} is the list of parameters, (ii)
\code{lidat} is the list containing the data, and (iii) \code{ctrl}
is a named list (see the help page of
\code{GeneralSingleMetropolis}). Note that the argument \code{ctrl}
is not used by the function here.}
}
\value{
the value of the DIC.
}
\description{
Deviance Information Criterion
}
\details{
Calculates the deviance information criterion of a model.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}}.
}

