% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{getParameterVector}
\alias{getParameterVector}
\alias{increaseBurnin}
\alias{nit}
\title{Various functions used to manipulate results of the Metropolis Algorithm}
\usage{
getParameterVector(resuParms, r)

nit(resuParms)

increaseBurnin(resuParms, newBurnin = attr(resuParms, "nburnin"))
}
\arguments{
\item{resuParms}{an object of class \code{GSMetrop}}

\item{r}{iteration to be extracted}

\item{newBurnin}{new size of the burnin sample (in number of
*performed* iterations: in other words, if the objects results from
a calculation which carried out 5000 iterations thinned every 5
iterations, with a 10 iterations burnin period, then the original
object will contain 990 stored iterations. Setting
\code{newBurnin=500} will remove the 500 first performed
iterations, i.e. the first (500 - 10 old burnin)/(thin=5) = 98
iterations stored in the object).}
}
\description{
Various functions used to manipulate results of the Metropolis Algorithm
}
\details{
\code{getParameterVector} extracts the parameter vector
corresponding to a given iteration.  \code{increaseBurnin} can be
used to increase a posteriori the size of the burnin
sample. \code{nit} returns the number of stored iterations in a
GSMetrop object.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}} for examples of use.
}

