% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{logit}
\alias{invlogit}
\alias{logit}
\title{transformations used in the package}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{parameter of a model.}
}
\description{
transformations used in the package
}
\details{
\code{logit} calculates the logit of the argument, and
\code{invlogit} calculates the inverse logit transform.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}} for examples of use.
}

