% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{package}
\name{metroponcfs}
\alias{metroponcfs}
\alias{metroponcfs-package}
\title{metroponcfs: A package for implementing the Metropolis algorithm}
\description{
The metroponcfs provides many interesting functions for
implementing the Metropolis algorithm, but the two main help pages
of interest are \code{GeneralSingleMetropolis}, which implements
and demonstrate how to use the Metropolis algorithm, and
\code{c.GSMetrop}, which demonstrate how to combine different
chains in an object.
}

