% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{param}
\alias{param}
\title{Extracts the name of the parameters from a model}
\usage{
param(x)
}
\arguments{
\item{x}{an object of class \code{CMM} or \code{GSMetrop}}
}
\value{
a character vector giving the names of the parameters.
}
\description{
Extracts the name of the parameters from a model
}
\details{
Extracts the name of the parameters from a \code{CMM} or
\code{GSMetrop} object
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples of use.
}

