% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{plot.CMM}
\alias{plot.CMM}
\title{graphical display of the result of several Metropolis runs}
\usage{
\method{plot}{CMM}(x, which = names(x[1]), show = 100,
  col = rainbow(length(x)), ...)
}
\arguments{
\item{x}{an object of class \code{CMM}}

\item{which}{vector of name of parameters to be displayed}

\item{show}{the number of plot to show on the graphical device}

\item{col}{a character vector giving the colors of the different chains}

\item{\dots}{additional arguments to be passed to and from other
functions}
}
\description{
graphical display of the result of several Metropolis runs
}
\details{
plot the mixing of several metropolis runs
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples of use.
}

