% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{plot.GSMetrop}
\alias{plot.GSMetrop}
\title{graphical display of the result of a Metropolis run}
\usage{
\method{plot}{GSMetrop}(x, which = names(x), show = 100, ...)
}
\arguments{
\item{x}{an object of class \code{GSMetrop}}

\item{which}{vector of name of parameters to be displayed}

\item{show}{the number of plot to show on the graphical device}

\item{\dots}{additional arguments to be passed to and from other
functions}
}
\description{
graphical display of the result of a Metropolis run
}
\details{
plot the mixing of a metropolis run
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}} for examples of use.
}

