% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{print.CMM}
\alias{print.CMM}
\title{Prints a CMM object}
\usage{
\method{print}{CMM}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CMM}}

\item{\dots}{additional arguments to be passed to other functions.}
}
\description{
Prints a CMM object
}
\details{
\code{print.CMM} function displays a short summary of an
object of class \code{CMM} (combined multiple Metropolis).
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples
}

