% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{recteal}
\alias{recteal}
\title{Common Teal captured and Ringed in Abberton Reservoir, UK,
and Camargue, France.}
\format{A list with the following elements:

\describe{

\item{recaptures}{this element is a data.frame containing the
following information for each recapture of common teal: (i) a
variable named \code{date}, storing the number of years elapsed
between the initial capture and recapture, (ii) a variable named
\code{Abberton}, indicating whether the recaptured animal was
initially captured at Abberton Reservoir, UK (=1) or in Camargue,
southern France (=0), and (iii) two variables named \code{x} and
\code{y} containing the coordinates of the recapture (coordinate
system: Lambert azimuthal equal-area).}

\item{rotationMatrix}{this 2x2 matrix M contains the two vectors
\code{cbind(m1, m2)} used by Guillemain et al. to rotate the
geographical coordinates in a new coordinate system (see Guillemain
et al.). Thus, if \code{x} is a vector of length two containing the
x and y coordinates of a location in the Lambert azimuthal
equal-area system, \code{z <- x\%*\%M} contains the coordinates of
this point in this new system.}

\item{inverseRotationMatrix}{this 2x2 matrix \code{R} allows to
transform the coordinates of a point from the new coordinate system
to the old one. Thus, if \code{z} is a vector of length two
containing the coordinates of a point in the new coordinate system,
\code{x <- z\%*\%R} contains the coordinates of this point in the
original Lambert azimuthal equal-area system.}

\item{knots}{this vector contains the coordinates of
the 26 knots in the new coordinate system used to define the
B-spline basis in the paper of Guillemain et al.}

\item{lipum}{a list containing the parameters of the updating
mechanisms used in the Metropolis algorithm.}
}}
\usage{
recteal
}
\description{
Common Teal captured and Ringed in Abberton Reservoir, UK,
and Camargue, France.
}
\details{
Dataset describing the recapture locations of common teals
initially captured and ringed in two Western European places: (i)
in Abberton Reservoir, UK, and (ii) in Camargue, sourthern
France. See Guillemain et al. for a complete description of this
dataset.  To circumvent copyright issues, we provide here an
altered version of the dataset used by these authors: we selected a
random sample of 75\% of the recaptures of the original data,
keeping only the location information (i.e. only the x and y
coordinates of the recaptures), and we added a random noise to
these locations (we moved every bird recapture location randomly by
a distance comprised between 0 and 100 km).
}
\references{
Guillemain M., Calenge C., Champagnon J. and Hearn
R. in prep. Determining the boundaries and plasticity of migratory
bird flyways: a Bayesian model for Common Teal Anas crecca
in Western Europe.
}
\keyword{datasets}

