% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{rectealmodeltime}
\alias{rectealmodeltime}
\title{Bayesian Model Fitted to Estimate the Boundaries of Common
Teal Flyways (time effect)}
\format{An object of class \code{CMM} (see
\code{help("c.GSMetrop")} for additional details on objects of
class \code{CMM}.}
\usage{
rectealmodeltime
}
\description{
Bayesian Model Fitted to Estimate the Boundaries of Common
Teal Flyways (time effect)
}
\details{
This model is fully described in the vignette "flyways".
Please type \code{vignette("flyways")} for more details.
}
\references{
Guillemain M., Calenge C., Champagnon J. and Hearn
R. in prep. Determining the boundaries and plasticity of migratory
bird flyways: a Bayesian model for Common Teal Anas crecca
in Western Europe.
}
\keyword{datasets}

