% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{sensitivityAngles}
\alias{sensitivityAngles}
\title{Results of the Sensitivity Analysis for the Rotation Angles}
\format{A list with three elements named \code{fit} (containing the
model fits), \code{proportion} and \code{differenceWithRef}.}
\usage{
sensitivityAngles
}
\description{
Results of the Sensitivity Analysis for the Rotation Angles
}
\details{
This model is fully described in the vignette "flyways".
Please type \code{vignette("flyways")} for more details.
}
\references{
Guillemain M., Calenge C., Champagnon J. and Hearn
R. in prep. Determining the boundaries and plasticity of migratory
bird flyways: a Bayesian model for Common Teal Anas crecca
in Western Europe.
}
\keyword{datasets}

