% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{singleStep}
\alias{MultiNormalStep}
\alias{MultipleIndependentSteps}
\alias{singleStep}
\title{Updating mechanisms used in the metropolis algorithm}
\usage{
singleStep(par, lidat, nam, pum, ctrl, logposterior, postcur, optimiser, debug,
  dumpfile, composition)

MultipleIndependentSteps(par, lidat, nam, pum, ctrl, logposterior, postcur,
  optimiser, debug, dumpfile, composition)

MultiNormalStep(par, lidat, nam, pum, ctrl, logposterior, postcur, optimiser,
  debug, dumpfile, composition)
}
\arguments{
\item{par}{a named list containing the parameters of the model (the
names of the list correspond to the names of the parameters)}

\item{lidat}{a named list containing the variables and constants
required by the model (especially the function \code{logposterior}}

\item{nam}{character string. The name of the parameter to be
updated.}

\item{pum}{for \code{singleStep}, the standard deviation of the
Gaussian distribution used to generate the proposal. For
\code{MultipleIndependentSteps}, a vector containing the standard
deviations for the Gaussian distribution used to generate the
proposal of every element of the vector. For \code{MultiNormalStep},
the covariance matrix of the multinormal distribution.}

\item{ctrl}{a named list with two parameters: (i) \code{namePar} is
the name of the parameter in \code{par} that is updated, (ii) if
this parameter is a vector where each component is updated in turn
(e.g. for \code{MultipleIndependentSteps}), \code{iter} indicates
the element of this vector that is updated (this element is ignored
in other cases).}

\item{logposterior}{a function used to calculate the log-posterior
probability of \code{par} (up to a constant).  This function must
have two arguments: \code{par} is the list of parameters and
\code{lidat} is the list containing the data used in the model.}

\item{postcur}{the current value of the posterior, before updating.}

\item{optimiser}{logical. If \code{TRUE}, the log-posterior
calculation is optimized for the current parameter (i.e. the
posterior is proportional to the log-posterior but the
proportionality constant is not necessarily the same for other
optimized parameters, so that the current posterior should be
calculated before every step).  If \code{FALSE} the function
\code{logposterior} returns the exact log-posterior.  In other
words, if \code{optimiser} is TRUE, the function \code{singleStep},
etc. calculates the current posterior (prior to updating) and does
not take into account the value of \code{postcur} passed to the
function. If FALSE, the function relies on the value of
\code{postcur} passed as argument for the updating.}

\item{debug}{logical. If \code{TRUE} the updating should be run in
debugging mode.}

\item{dumpfile}{character. The name of the debugging file.}

\item{composition}{character. How composition of parameters are
managed (see the help page of
\code{GeneralSingleMetropolis}). Currently, only used by
\code{MultipleIndependentSteps}.}
}
\value{
A list with three elements: (i) parm is the updated vector
of parameters, (ii) postcur is the value of the (optionally
optimized) posterior after updating, and (iii) accept is a vector
containing TRUE if an element proposed has been accepted and FALSE
otherwise
}
\description{
Updating mechanisms used in the metropolis algorithm
}
\details{
\code{singleStep} implements an updating mechanism for a
unique parameter based on a normal distribution.
\code{MultipleIndependentSteps} implements the same updating
mechanism for a vector of parameters. \code{MultiNormalStep}
implements an updating mechanism relying on the sampling of a
multinormal distribution. These functions are used by the package,
but are not to be used directly by the user.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}}
}

