% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{[.CMM}
\alias{[.CMM}
\title{Extract elements from a CMM object}
\usage{
\method{[}{CMM}(x, i)
}
\arguments{
\item{x}{an object of class \code{CMM}}

\item{i}{a vector of positive integer values used as indices of the
elements to extract.}
}
\value{
an object of class "CMM".  If only one object is passed,
the function returns the input object without transformation.
}
\description{
Extract elements from a CMM object
}
\details{
\code{Extract} function can be used to extract one or
several elements from an object of class \code{CMM} (combined
multiple Metropolis).
}
\note{
negative indices (for element removal) are not allowed.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples
}

