% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{summary.CMM}
\alias{summary.CMM}
\title{Summary information on a 'CMM' object}
\usage{
\method{summary}{CMM}(object, which = names(object[1]), perchain = TRUE,
  ...)
}
\arguments{
\item{object}{an object of class \code{CMM}}

\item{which}{the name of the parameters to summarize}

\item{perchain}{logical. whether the statistics should be calculated for each chain separately or globally}

\item{\dots}{additional parameters to be passed}
}
\value{
a data.frame with the required statistics.
}
\description{
Summary information on a 'CMM' object
}
\details{
Calculate mean and SD of parameters from a
'CMM' object
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples of use.
}

