% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{summary.GSMetrop}
\alias{summary.GSMetrop}
\title{Summary information on a 'GSMetrop' object}
\usage{
\method{summary}{GSMetrop}(object, which = names(object), accept = TRUE,
  parameters = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{GSMetrop}}

\item{which}{the name of the parameters to summarize}

\item{accept}{logical. whether to calculate acceptation rates}

\item{parameters}{logical. whether to calculate mean and sd}

\item{\dots}{additional parameters to be passed}
}
\description{
Summary information on a 'GSMetrop' object
}
\details{
Calculate acceptation rates, mean and SD from a
'GSMetrop' object
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{GeneralSingleMetropolis}} for examples of use.
}

