% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{tocoda}
\alias{tocoda}
\title{Exports object toward the coda package}
\usage{
tocoda(x, which = param(x))
}
\arguments{
\item{x}{an object of class \code{CMM} or \code{GSMetrop}}

\item{which}{a character vector indicating which parameters should be exported}
}
\value{
if \code{x} is an object of class GSMetrop, an object of
class \code{mcmc}. If \code{x} is an object of class CMM, an object
of class {mcmc.list}
}
\description{
Exports object toward the coda package
}
\details{
\code{tocoda} exports objects created with the package
metroponcfs toward classes suitable for further analysis with the
coda package.
}
\author{
Clement Calenge, \email{clement.calenge@oncfs.gouv.fr}
}
\seealso{
\code{\link{c.GSMetrop}} for examples of use.
}

